Refactored code into folders and redid all imports(lots of imports)
