+++ frontend/src/components/pages/CardDetails/Presentation.js
