+++ frontend/src/components/pages/CardDetails/redux/index.js
