+++ frontend/src/components/pages/CardDetails/redux/operations.js
