+++ frontend/src/components/pages/CardDetails/redux/reducers.js
