+++ frontend/src/components/pages/CardDetails/redux/types.js
